requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
mediapipe>=0.10.31,<0.11
opencv-python==4.9.0.80
numpy==1.26.3
pyyaml==6.0.1
pytest==7.4.4
ruff==0.2.1
